# Optimistically resume partitions early