--- a/zio-kafka-test/src/test/scala/zio/kafka/Benchmarks.scala
+++ b/zio-kafka-test/src/test/scala/zio/kafka/Benchmarks.scala
@@ -82,7 +82,6 @@ object ZIOKafka extends ZIOAppDefault {
       .withProperty(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
       .withProperty("fetch.min.bytes", "128000")
       .withPollTimeout(50.millis)
-      .withPerPartitionChunkPrefetch(4)
 
     (Console.readLine *>
       Clock