--- a/zio-kafka/src/main/scala/zio/kafka/consumer/ConsumerSettings.scala
+++ b/zio-kafka/src/main/scala/zio/kafka/consumer/ConsumerSettings.scala
@@ -10,7 +10,6 @@ import zio.kafka.security.KafkaCredentialStore
  * @param properties
  * @param closeTimeout
  * @param pollTimeout
- * @param perPartitionChunkPrefetch
  * @param offsetRetrieval
  * @param rebalanceListener
  * @param restartStreamOnRebalancing
@@ -27,7 +26,6 @@ case class ConsumerSettings(
   properties: Map[String, AnyRef],
   closeTimeout: Duration,
   pollTimeout: Duration,
-  perPartitionChunkPrefetch: Int,
   offsetRetrieval: OffsetRetrieval = OffsetRetrieval.Auto(),
   rebalanceListener: RebalanceListener = RebalanceListener.noop,
   restartStreamOnRebalancing: Boolean = false,
@@ -65,9 +63,6 @@ case class ConsumerSettings(
   def withOffsetRetrieval(retrieval: OffsetRetrieval): ConsumerSettings =
     copy(offsetRetrieval = retrieval)
 
-  def withPerPartitionChunkPrefetch(prefetch: Int): ConsumerSettings =
-    copy(perPartitionChunkPrefetch = prefetch)
-
   def withPollTimeout(timeout: Duration): ConsumerSettings =
     copy(pollTimeout = timeout)
 
@@ -102,7 +97,6 @@ object ConsumerSettings {
       properties = Map.empty,
       closeTimeout = 30.seconds,
       pollTimeout = 50.millis,
-      perPartitionChunkPrefetch = 2,
       offsetRetrieval = OffsetRetrieval.Auto(),
       runloopTimeout = defaultRunloopTimeout
     )