Optimistically resume partitions that are likely to need more data before the next poll.

Solves #800.

Benchmark reports:

master:
```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   465.958 ± 100.053  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   433.205 ±  85.996  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   664.730 ±  12.088  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1999.502 ± 717.557  ms/op
```

This branch (commit https://github.com/zio/zio-kafka/pull/809/commits/379b490c8144e077cbac7361cbb6e176af5feae5):
```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   226.624 ±  19.807  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   316.707 ±  29.479  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   654.995 ±  10.792  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1132.004 ± 174.654  ms/op
```

1.8 times faster!

After removing pre-fetching (commit https://github.com/zio/zio-kafka/pull/809/commits/ca4e88cb536026b3ddd88afa9810481b89112fc5):

```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   298.363 ±  57.858  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   300.736 ±  23.548  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   654.316 ±  15.199  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1002.338 ± 101.240  ms/op
```

2 times faster! 🚀