@erikvanoosten 2023-04-13
Benchmark reports:

master:
```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   465.958 ± 100.053  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   433.205 ±  85.996  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   664.730 ±  12.088  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1999.502 ± 717.557  ms/op
```

This branch:
```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   396.172 ±  68.890  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   407.719 ±  52.321  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   746.998 ±  28.032  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  2029.001 ± 538.249  ms/op
```

All lower numbers but also all within the error-margin.