@erikvanoosten 2023-04-14
Patterns fixed and benchmark updated.