@erikvanoosten 2023-04-14
The earlier benchmark contains a feedback loop that does a lot more resuming than it should.

With the latest change we get:

```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   524.119 ±  85.142  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   415.531 ±  65.929  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   666.154 ±  17.329  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1793.227 ± 635.461  ms/op
```

It is faster but all in the error margin. This is not very encouraging.