@erikvanoosten 2023-04-14
Yes, this is better:

```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25   240.436 ±  23.397  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25   324.053 ±  28.001  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25   660.852 ±  10.127  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  1197.165 ± 223.556  ms/op
```