@svroonland 2023-04-14
Ran a comparison again on my machine (2019 2,4 GHz 8-Core Intel Core i9)

This branch:

```
[info] Benchmark                                  Mode  Cnt    Score     Error  Units
[info] ConsumerBenchmark.throughput               avgt   25  189,833 ±   7,312  ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25  287,570 ±   2,914  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25  676,153 ±  21,006  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  863,424 ± 122,444  ms/op
```

#803 
```
[info] Benchmark                                  Mode  Cnt    Score    Error  Units
[info] ConsumerBenchmark.throughput               avgt   25  189,069 ± 6,276   ms/op
[info] ConsumerBenchmark.throughputWithCommits    avgt   25  289,007 ± 3,590   ms/op
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   25  693,354 ± 31,611  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   25  838,962 ± 70,050  ms/op
```

I'd say performance is virtually equal. I do think #803 is simpler to understand and explain..