@erikvanoosten 2023-04-14
> I'd say performance is virtually equal. I do think #803 is simpler to understand and explain..

I am happy about that because I found it hard to explain why one would be faster than the other.

I do not find this an argument given that Runloop is already pretty complicated. For the case you disagree, I think both have about equal complexity, perhaps 803 is a tiny bit more complex because it still needs the Request mechanism while 809 embraces it.

Anyways, IMHO we should look at memory usage. My hypothesis is that 809 can reach the same performance with less memory usage. I hope we can test this with the benchmarks as well.

@guizmaii is this command suitable for comparing memory usage as well?
```
zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 5 -foe true -prof async:event=alloc
```