@svroonland 2023-04-26
@erikvanoosten A few thoughts:

1. There is some initial learning period for this feature to 'kick in' right? So for the second Request of a partition we would not have yet prefetched the data, only for the third request we would 
2. Let's say we get a Request every `N` poll iterations. We will prefetching for that partition at iteration `N-1`. The Request itself may arrive anytime during this poll iteration and will have to wait until the poll completes. That would add some unneeded (in the sense that we could optimize it away) latency.
3. We're not guaranteed to get data for every partition that we resume, right? So we may still sometimes not be on time for the Request. 

That being said, this method improves benchmarks by a lot and more than #803 does and it's a great feature.