@erikvanoosten 2023-04-26
@svroonland 

>     1. There is some initial learning period for this feature to 'kick in' right? So for the second Request of a partition we would not have yet prefetched the data, only for the third request we would

Yes, that is correct. Currently the shortest pattern is 3 polls, so 3 polls (150ms with default settings) is the lowest learning time. After the first 3 polls more pre-fatch patterns can match so more polls become eligible for pre-fetching.

>     2. Let's say we get a Request every `N` poll iterations. We will prefetching for that partition at iteration `N-1`. The Request itself may arrive anytime during this poll iteration and will have to wait until the poll completes. That would add some unneeded (in the sense that we could optimize it away) latency.

Correct. However, since the previous approach had the same issue, a pre-fetch still reduces the average waiting time with 1 full poll cycle.

>     3. We're not guaranteed to get data for every partition that we resume, right? So we may still sometimes not be on time for the Request.

Correct. Same is true for previous approach. So a pre-fetch still lowers the average waiting time with 1 poll cycle.

The algorithm will also sometimes predict a pre-fetch when it is not yet necessary (reducing waiting time though waiting for the next poll would have been better), or fail to predict one when it is necessary (increasing wait time). Combined this somewhat lowers the effectiveness of the current algorithm. More fancy algorithms should be possible.

For a more accurate estimate of the waiting time reduction, you can probably iterate all possibilities or do a montecarlo simulation for different situations. My math skills are not up to that task though 😄 

I was thinking about adding some metrics to see how well the algorithm is doing. However, we can only detect failure to predict pre-fetches. We cannot detect whether a predicted pre-fetch was correct (though we could when we track more information). To make meaningful statements about the pre-fetching algorithm we need both metrics.