@guizmaii 2023-04-26
Performances measured on my computer:
```scala
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.kafkaClients":
[info]   641.297 ±(99.9%) 2.210 ms/op [Average]
[info]   (min, avg, max) = (634.237, 641.297, 659.168), stdev = 4.465
[info]   CI (99.9%): [639.087, 643.508] (assumes normal distribution)
...
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.zioKafka":
[info]   734.035 ±(99.9%) 18.438 ms/op [Average]
[info]   (min, avg, max) = (674.275, 734.035, 875.752), stdev = 37.247
[info]   CI (99.9%): [715.596, 752.473] (assumes normal distribution)
...
[info] Benchmark                                  Mode  Cnt    Score    Error  Units
[info] ConsumersComparisonBenchmark.kafkaClients  avgt   50  641.297 ±  2.210  ms/op
[info] ConsumersComparisonBenchmark.zioKafka      avgt   50  734.035 ± 18.438  ms/op
```